update webpack setting & fix test

+++ server/client-entry.js
@@ -1,18 +1,32 @@
 import {join, relative, resolve} from 'path';
 import App from './app';
 import {outputCommonJSDir, clientTmpEntryDir} from './CONFIG';
+import {isFile} from './utils';
+import {isString, isArray} from 'lodash';
 import fs from 'fs';
 import mkdirp from 'mkdirp';
 
 exports.addClientEntry = (dir, config) => {
   const {entry} = config;
   const clientEntryDir = clientTmpEntryDir(dir);
-  mkdirp.sync(clientEntryDir);
   config.clientEntry = {};
   for (let key in entry) {
-    const clientEntryPath = join(clientEntryDir, entry[key]);
+    const entryPath = entry[key];
+    let clientEntryPath;
+    if (isString(entryPath)) {
+      clientEntryPath = join(clientEntryDir, entryPath);
+    } else if (isArray(entryPath)) {
+      clientEntryPath = entryPath.map(path => {
+        // if isFile: transform its path to .coren file, else: is node_module
+        if (isFile(path)) {
+          return join(clientEntryDir, path);
+        }
+        return path;
+      });
+    }
     config.clientEntry[key] = clientEntryPath;
   }
+
   return config;
 };
 
@@ -45,20 +59,28 @@ exports.createClientTmpEntryFile = (dir, config) => {
     , document.getElementById('root'));`;
   for (let key in entry) {
     // entryPath: ./src/index.js
-    const entryPath = entry[key];
-    // clientTmpDir: /.coren/tmp
-    const clientTmpDir = clientTmpEntryDir(dir);
-    // clientEntryPath: /.coren/src/index.js
-    const clientEntryPath = join(clientTmpDir, entryPath);
-    // clientEntryDir: /.coren/src
-    const clientEntryDir = resolve(clientEntryPath, '../');
-    mkdirp.sync(clientEntryDir);
-    // importPath
-    const importPath = relative(clientEntryDir, join(dir, entryPath));
-    const tmpJS =
-      `${clientImport}
-       import App from "${importPath}";
-       ${clientRender}`;
-    fs.writeFileSync(clientEntryPath, tmpJS, 'utf8');
+    let entryPath = entry[key];
+    if (isString[entryPath]) {
+      entryPath = [entryPath];
+    }
+    for (let i in entryPath) {
+      const path = entryPath[i];
+      if (isFile(path)) {
+        // clientTmpDir: /.coren/tmp
+        const clientTmpDir = clientTmpEntryDir(dir);
+        // clientEntryPath: /.coren/tmp/src/index.js
+        const clientEntryPath = join(clientTmpDir, entryPath);
+        // clientEntryDir: /.coren/tmp/src
+        const clientEntryDir = resolve(clientEntryPath, '../');
+        mkdirp.sync(clientEntryDir);
+        // importPath
+        const importPath = relative(clientEntryDir, join(dir, entryPath));
+        const tmpJS =
+          `${clientImport}
+           import App from "${importPath}";
+           ${clientRender}`;
+        fs.writeFileSync(clientEntryPath, tmpJS, 'utf8');
+      }
+    }
   }
 };
